% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkbehaviour.R
\name{checkbehaviour}
\alias{checkbehaviour}
\title{data checks}
\usage{
checkbehaviour(b.source)
}
\arguments{
\item{b.source}{data frame with behavioural data}
}
\description{
check interaction data for inconsistencies
}
\examples{
# nothing yet
}
