% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpresence.R
\name{checkpresence}
\alias{checkpresence}
\title{Title}
\usage{
checkpresence(presence, b.source = NULL, ot.source = NULL)
}
\arguments{
\item{presence}{presence "matrix" (usually a data frame...); see \code{\link{createnullpresence}}. \code{presence2} is only necessary if focal-nonfocal dyads are to be considered at the same time, e.g. when interested in male-male focal dyads and male-female focal-nonfocal dyads. In this case, \code{presence2} would be the female presence data.}

\item{b.source}{data frame with behavioural data}

\item{ot.source}{data frame with observation time}
}
\value{
a matrix, with dates and roles (actor, etc...) at which presence and behavioural data don't match. Same for observation time data. In case both are specified, a list is returned.
}
\description{
Title
}
\examples{
\dontrun{
data(dataset3)
checkpresence(dataset3$pres, dataset3$dataseq)
checkpresence(dataset3$pres, dataset3$dataseq, dataset3$ot)

x <- datagen(no=100, ni=7, presence=c(0.8,0.4))
checkpresence(x$pres, x$dataseq, x$ot)
}
}
