% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cores.R
\name{cores}
\alias{cores}
\title{coresidency of a dyad}
\usage{
cores(ID1, ID2, daterange, presence, outp, pr.dates = NULL)
}
\arguments{
\item{ID1}{character, IDs}

\item{ID2}{character, IDs}

\item{daterange}{character of length 2, e.g. \code{c("2010-01-01", "2010-04-04")}}

\item{presence}{character: name of presence data frame (i.e. in quotes), or matrix without date column. in the latter case make sure that \code{pr.dates=} is supplied}

\item{outp}{\code{"abs"} (returning number of days), or \code{"prop"} (proportion of days given the date range), or \code{"2days"} (both IDs present on both dates), see details}

\item{pr.dates}{optional character vector which corresponds to the \code{presence=} matrix. Should speed up computations in randomization tests if presence of two individuals on two specific days is needed.}
}
\value{
numeric, or logical
}
\description{
coresidency of a dyad
}
\details{
the \code{outp="2days"} option is useful only in the context of randomization procedures, and as such probably not useful for normal applications of the function.

if the \code{pr.dates=} argument is used this implies that the actual presence matrix is a true numeric matrix while the corresponding dates are in a separate vector (of the same length as the number of rows of the presence data set). So, it's probably also an argument that is not useful for the typical use.
}
\examples{
# load an example data set:
data(dataset3)
# extract the presence table from it:
pres <- dataset3$presence
head(pres)
cores("a", "d", daterange = c("2000-01-01", "2000-01-06"), presence = "pres", outp = "abs")
cores("a", "d", daterange = c("2000-05-01", "2000-06-01"), presence = "pres", outp = "prop")
}
