% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createnullot.R
\name{createnullot}
\alias{createnullot}
\title{create null data}
\usage{
createnullot(IDs, from, to)
}
\arguments{
\item{IDs}{character, IDs}

\item{from}{start date}

\item{to}{end date}
}
\value{
data.frame, i.e. observation time table
}
\description{
create null data
}
\examples{
createnullot(letters[1:5], "2000-01-01", "2000-01-05")
}
