% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{datasets}
\alias{datasets}
\alias{dataset1}
\alias{dataset2}
\alias{dataset3}
\alias{dataset4}
\alias{domseq4}
\title{example data sets for socialindices}
\format{
\itemize{
\item \code{dataseq} data.frame with interaction data
\item \code{ot} data.frame with observation time of individual protocols
\item \code{presence} presence table of focal individuals
\item \code{presence2} presence table of nonfocal individuals
\item \code{prefdyads} dyads with higher rate/longer duration of interactions
\item \code{focals} character, focal individuals
\item \code{nonfocs} character, nonfocal individuals
}
}
\usage{
dataset1
}
\description{
Fictious examples of data sets to calculate social bonds
}
\details{
The data sets included have been generated with the \code{\link{datagen}} function. See examples for the code...
}
\examples{
\dontrun{
dataset1 <- datagen(no=2000, ni=10, nadd=20, friendprop=0.2, presence=c(0.5, 0.4))
dataset2 <- datagen(no=2000, ni=30, nadd=0,  friendprop=0.2, presence=c(0.5, 0.4))
dataset3 <- datagen(no=200,  ni=3,  nadd=3,  friendprop=0.2, presence=c(0.5, 0.4))
dataset4 <- datagen(1000, ni=14)
domseq4 <- domgen(dataset4, 500, drawprop=0)
}
}
\keyword{datasets}
