% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dolphins.R
\docType{data}
\name{dolphins}
\alias{dolphins}
\title{dolphins in Porpoise Bay in 1996-1997}
\format{data.frame, 18 indidivuals A through R, 0 indicates individual was absent, 1 indicates individual was present}
\usage{
dolphins
}
\description{
Sightings of identifiable Hector's dolphins in Porpoise Bay in 1996-1997 for 40 groups
}
\references{
Bejder, L., Fletcher, D., & Bräger, S. (1998). A method for testing association patterns of social animals. Animal Behaviour, 56(3), 719-725. \url{http://dx.doi.org/10.1006/anbe.1998.0802}
}
\keyword{datasets}
