% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domgen.R
\name{domgen}
\alias{domgen}
\title{generate random data for bond package}
\usage{
domgen(bondsdata = NULL, no = 100, ago = c("suppl", "thrt", "fight"),
  agow = c(0.6, 0.3, 0.1), drawprop = 0.1, reversals = 0.1)
}
\arguments{
\item{bondsdata}{output from \code{\link{datagen}}}

\item{no}{number of interactions to be created}

\item{ago}{character, determines behaviours for the "agonistic" interactions of \code{domgen}}

\item{agow}{numeric, weights of occurences of behaviours in \code{domgen}}

\item{drawprop}{proportion of interactions that end undecided}

\item{reversals}{proportion of interactions that end against the implicit rule regarding which individual wins. By default this is alphabetic, i.e. of two individuals interacting, the winner precedes the loser in alphabetic order}
}
\value{
a data.frame with winner/loser interactions
}
\description{
generate random data for bond package
}
\details{
\code{domgen()} does not yet work by itself, i.e. it relies on specifying \code{bondsdata} being the output from \code{\link{datagen}}.
}
\examples{
x <- datagen(no=20, ni=3)
# observation time table
x$ot
# interactions
x$dataseq

# and for dominance interactions
y <- domgen(x)
}
