% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipdyads.R
\name{.flipdyads}
\alias{.flipdyads}
\title{flip dyads in (0,1) table}
\usage{
.flipdyads(x, x.dates, flips, presence = NULL, pr.dates = NULL, x.fo = NULL)
}
\arguments{
\item{x}{data frame with association data}

\item{x.dates}{Date, dates corresponding to rows in \code{x} (\code{x.dates=}), i.e. of same length as rows of \code{x}, and corresponding to \code{presence} (\code{pr.dates=})}

\item{flips}{numeric, how many flips to perform}

\item{presence}{presence matrix; see \code{\link{createnullpresence}}}

\item{pr.dates}{Date, dates corresponding to rows in \code{x} (\code{x.dates=}), i.e. of same length as rows of \code{x}, and corresponding to \code{presence} (\code{pr.dates=})}

\item{x.fo}{character string corresponding to rows in \code{x} with focal animal IDs, if supplied, focal animals are excluded from flipping, i.e. focals remain in their party/association}
}
\value{
a table with the same dimensions as \code{x}, except that the date column is removed if there was one in \code{x} when it was supplied to the function.
}
\description{
flip dyads in 0/1 matrix to randomize data set
}
\examples{
\dontrun{
data(dolphins)
# create a presence table (all being present all the time)
dpres <- createnullpresence(colnames(x)[2:19], from="2000-01-01", "2000-02-09")
# without presence
colSums(flipdyads(dolphins, 10))
# with presence
colSums(flipdyads(dolphins, 10, dolphins$Date, presence))

dolphins[,-1] - flipdyads(dolphins, 10)
sum(abs(dolphins[,-1] - flipdyads(dolphins, 20)))
}
}
\references{
Whitehead 2008; Bejder et al 1998
}
