% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot.R
\name{dot}
\alias{dot}
\title{dyadic observation time}
\usage{
dot(ot.source, dyads, daterange = NULL, presence = NULL,
  ot.style = "table", exclcols = NULL)
}
\arguments{
\item{ot.source}{observation time data, for the moment only supported in table form, not yet in party-style. Must contain a "focal" column, a date column, and a observation time column}

\item{dyads}{a character matrix with four columns, the first two correspond to the individual ids, and the third and fourth are dyadic, first in alphabetical order, second the reversed of that, and in which individuals are separated by \code{"_@_"}. See \code{\link{makedyads}}. If only a character vector is used, dyads are generated by \code{makedyads}}

\item{daterange}{character of length 2, with date range...}

\item{presence}{presence "matrix" (usually a data frame...); see \code{\link{createnullpresence}}. \code{presence2} is only necessary if focal-nonfocal dyads are to be considered at the same time, e.g. when interested in male-male focal dyads and male-female focal-nonfocal dyads. In this case, \code{presence2} would be the female presence data.}

\item{ot.style}{character, either \code{"table"} (default) or \code{"party"}, ie. party style...}

\item{exclcols}{character string with column names (or numerical index of these) to be excluded. Normally used to exclude columns that neither represent IDs, date or focal.}
}
\value{
data.frame
}
\description{
dyadic observation time
}
\details{
returned are only dyads that were co-resident at least one day and were observed for more than zero minutes/hours (depending on the unit of the OT) MAYBE NOT TRUE ANYMORE???
}
\examples{
\dontrun{
data(dataset3)
# limit presence to focals (not yet differentiated between presence and presence2
in the example data sets..)
focalpresence <- dataset3$pres[, c("date", "m", "g", "y")]
dot(daterange = c("2000-01-11", "2000-04-01"), OT = dataset3$ot, presence = dataset3$presence)
dot(daterange = c("2000-01-11", "2000-04-01"), OT = dataset3$ot, presence = NULL)
}
}
