% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friends.R
\name{friends}
\alias{friends}
\title{top partners}
\usage{
friends(
  x,
  criterion = NULL,
  indexcols = "DSI",
  IDcols = c("i1", "i2"),
  tableout = FALSE
)
}
\arguments{
\item{x}{output from \code{\link{DSI}}}

\item{criterion}{either one of \code{"1sd"}, \code{"2sd"}, \code{"10perc"}, \code{"5perc"}, or numeric}

\item{indexcols}{character, the column that contains the index of interest}

\item{IDcols}{character of length 2, column names of the two dyad members}

\item{tableout}{logical by default \code{FALSE}, if \code{TRUE} and a group level criterion was selected (not numeric), a table with the dyads in descending order of their dyadic index is returned}
}
\value{
a list
}
\description{
top partners
}
\details{
if the criterion is one of \code{"1sd"}, \code{"2sd"}, \code{"10perc"}, \code{"5perc"} the cutoff is considered across all dyads in the data set. If the criterion is numeric, then the x top partners of each individual will be returned regardless of absolute values of the index. If \code{criterion=NULL}, its default, all possible dyads will be returned sorted by size within individual.
}
\examples{
data(dataset3)
dsi <- DSI(dataset3$dataseq, presence=dataset3$presence, ot.source=dataset3$ot,
onlyfocaldyads = FALSE)
friends(dsi, "2sd")
friends(dsi, "2sd", tableout = TRUE)
friends(dsi, 3)
}
