% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limit2daterange.R
\name{limit2daterange}
\alias{limit2daterange}
\title{limit dataset to date range}
\usage{
limit2daterange(x, daterange, omitdatecol = FALSE)
}
\arguments{
\item{x}{data.frame with a date column in it \code{YYYY-MM-DD}}

\item{daterange}{character of length 2, e.g. \code{c("2010-01-01", "2010-04-04")}}

\item{omitdatecol}{should the actual date column be omitted from the output, by default no (\code{FALSE})}
}
\value{
a subset of \code{x}
}
\description{
limit dataset to date range
}
\examples{
data(dataset3)
limit2daterange(dataset3$dataseq, c("2000-01-01", "2000-01-14"))
limit2daterange(dataset3$presence, c("2000-01-01", "2000-01-14"))
limit2daterange(dataset3$presence, c("2000-01-01", "2000-01-14"), TRUE)
}
