% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makedyads.R
\name{makedyads}
\alias{makedyads}
\title{dyads from single individual IDs}
\usage{
makedyads(IDs)
}
\arguments{
\item{IDs}{character, IDs}
}
\value{
a matrix with four columns, first two are the single IDs, third is the dyad in alphabetical order, fourth is dyad in reverse
}
\description{
dyads from single individual IDs
}
\examples{
makedyads(LETTERS[4:1])


}
