% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varname.R
\name{varname}
\alias{varname}
\title{find variable name in source data set}
\usage{
varname(type, src, return.na = FALSE, possibilities = FALSE)
}
\arguments{
\item{type}{character, one of \code{"date"}, \code{"time"}, \code{"focal"}, \code{"behaviour"}, \code{"actor"}, \code{"receiver"}, \code{"duration"}, \code{"obstime"} (observation time), \code{"partner"}, \code{"obsnumber"}, \code{"observer"} or \code{"IDs"}}

\item{src}{usually a data.frame}

\item{return.na}{by default \code{FALSE}, which results in the function stopping (returning an error) if a column is not found. If \code{TRUE} function will return \code{NA} in such a case (without producing an error). See also \code{Details}}

\item{possibilities}{logical. If \code{TRUE} returns the possible matches for the given \code{type} that are considered by the function}
}
\value{
character of variable name in the source data
}
\description{
detect variable/column names in source data set
}
\details{
the case of "duration" is special, such a column is not mandatory and therefore might not be found either because it does not exist or because its column name is not among the templates provided in the function. In case no duration column is found, it is returned as \code{NA} and a message is returned to inform the user about the possible ambiguity.
If \code{type="IDs"} an attempt is made to figure out whether there is one of the two possibilities of combinations of ID columns, i.e. (1) a focal and a partner column or (2) a focal, actor and receiver column
}
\examples{
x <- datagen()
varname("date", x[[1]])
varname("focal", x[[1]])

\dontrun{
  varname("observer", x[[1]]) # error
  varname("observer", x[[1]], return.na=TRUE) # works and returns NA
}

# return possible matches
varname("observer", x[[1]], possibilities=TRUE)
}
